+++ react-cors-spa/src/AddProduct.css
@@ -6,13 +6,3 @@
 .exitButton {
     margin: 10px;
 }
-
-.productdialogue {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
-
-input {
-    font-size: 1rem;
-}
