+++ react-cors-spa/src/Products.jsx
@@ -1,11 +1,14 @@
 import { useState, useCallback } from 'react';
 import ProductsTable from './ProductsTable';
 import AddProduct from './AddProduct';
+import ProductDialogue from './ProductDialogue';
+
 import Skeleton from '@mui/material/Skeleton';
 import { Box } from '@mui/material';
 
 function Products() {
     const [loading, setLoading] = useState(true);
+    const [dialogue, setDialogue] = useState(false);
 
     const handleLoading = useCallback(() => {
         setLoading(false);
@@ -16,8 +19,10 @@ function Products() {
             {loading ? 
             <Box className="addProductSkeleton">
                 <Skeleton variant="rounded" width={160} height={75} />
-            </Box> : <AddProduct />}
+            </Box>
+            : <AddProduct onDialogue={() => setDialogue(true)}/>}
             <ProductsTable onLoading={handleLoading} />
+            {dialogue && <ProductDialogue onDialogue={() => setDialogue(false)}/>}
         </Box>
     )
 }
